requirement.txt
# Core libraries
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
tensorflow==2.12.0
keras-rl2==1.0.5
gym==0.26.0
statsmodels==0.14.0

# Trading specific
backtrader==1.9.78.123
pyfolio==0.9.2
alpaca-trade-api==3.0.0
yfinance==0.2.28
ta-lib==0.4.26
vectorbt==0.24.3

# Data acquisition
polygon-api-client==1.12.3
websocket-client==1.6.1
requests==2.31.0
beautifulsoup4==4.12.2
finnhub-python==2.4.18

# Utilities
tqdm==4.66.1
joblib==1.3.1
python-dotenv==1.0.0